+++ CHANGELOG.md
@@ -6,8 +6,12 @@
   * [Mix] Automatically recompile on outdated Elixir version and show proper error messages
 
 * Deprecations
+  * [File] `File.stream_to!/3` is deprecated
   * [Kernel] `%` for sigils is deprecated in favor of `~`
   * [Kernel] `is_range/1` and `is_regex/1` are deprecated in favor of `Range.range?/1` and `Regex.regex?/1`
+  * [Stream] `Stream.after/1` is deprecated
+  * [URI] `URI.decode_query/1` is deprecated in favor of `URI.decode_query/2` with explicit dict argument
+  * [URI] Passing lists as key or values in `URI.encode_query/1` is deprecated
 
 * Backwards incompatible changes